Configured download data button
- For now, the CSV download contains all columns from Health Canada
data. When CFIA data comes in, may need to decided on what
columns to put into this download since CFIA data may have columns that
are not in Health Canada data and vice versa.

- Using the old CSV download function fails to download the raw data
since the data is too big that it takes up all the memory of the
browser. Instead, we should stream the download to users by
calling `URL.createObjectURL` that creates a new download object to be
streamed.
https://stackoverflow.com/questions/30167326/unable-to-download-large-data-using-javascript

WARNING: After we are done with the object from `URL.createObjectURL`,
remember to FREE UP the object by calling `URL.revokeObjectURL`
https://developer.mozilla.org/en-US/docs/Web/API/URL/revokeObjectURL_static